+++ Resources/config/core.xml
@@ -38,6 +38,16 @@
             <argument type="service" id="sonata.cache.manager" on-invalid="ignore" />
         </service>
 
+        <service id="sonata.block.templating.helper" class="Sonata\BlockBundle\Templating\Helper\BlockHelper" public="false">
+            <tag name="templating.helper" alias="sonata_block" />
+
+            <argument type="service" id="sonata.block.manager" />
+            <argument />
+            <argument type="service" id="sonata.block.renderer" />
+            <argument type="service" id="sonata.block.context_manager" />
+            <argument type="service" id="sonata.cache.manager" on-invalid="ignore" />
+        </service>
+
         <service id="sonata.block.loader.chain" class="Sonata\BlockBundle\Block\BlockLoaderChain">
             <argument />
         </service>
